WARNING: this breaks if you're running more than one dequeue on the same queue at the same time since they will both attempt to reshuffle the list! Is there a need to implement dequeue locking...?
So with that in mind, here is a proof of concept.

This version of Job.destroy uses RPOPLPUSH and a sentinel value to avoid loading the queue into client memory in the case with no job arguments. However, to avoid a race condition, Resque.pop needs to avoid popping the sentinel value or else Job.destroy will never give up! Therefore, Resque.pop has to check whether it is popping a sentinel value every time.

Also, there is a potential downside to all this: since the queue gets reshuffled while more jobs can be added and removed, the order of jobs can get very messed up... Perhaps there is a better solution than using RPOPLPUSH? Perhaps a LINDEX + sentinel would work.

This would all be a lot simpler if Redis's LREM could take a regular expression :).